Add linuxinstall scripts

+++ .bash_functions
@@ -14,5 +14,28 @@ find_file () {
 }
 
 create_bash_script() {
+    if [ $# -ne 1 ]; then
+        echo "Illegal number of arguments. This function requires a filename."
+        return 1
+    fi
+    if [ -e $1 ]
+    then
+        echo "file with name " $1 " already exists"
+        return 1
+    fi
     printf "#!/bin/bash\n\n\n" > $1 && chmod +x $1 && vim -c ":3j" -c "startinsert" $1
+}
+
+create_dockerfile() {
+    dockerfileName=$1
+    if [ -z $dockerfileName ]
+    then
+        dockerfileName="Dockerfile"
+    fi
+    if [ -e $dockerfileName ]
+    then
+        echo "file with name " $dockerfileName " already exists"
+        return 1
+    fi
+    printf "# vim:set ft=dockerfile:\n\n\n" > $dockerfileName && vim -c ":3j" -c "startinsert" $dockerfileName
 }
